reformat code base
