+++ web-app/app.py
@@ -3,14 +3,15 @@ from functools import wraps
 import os
 import requests
 import pymongo
-from flask import Flask, request, render_template, redirect, session
+from flask import Flask, request, render_template, redirect, session, jsonify
+from passlib.hash import pbkdf2_sha256
 
 app = Flask(__name__)
 # Connecting to local host and same db as ml's backend
 client = pymongo.MongoClient("mongodb://db:27017")
 db = client["Isomorphism"]
 
-# Login require decorators
+# Utilities
 def login_required(f):
     """login required decorators"""
     @wraps(f)
@@ -21,6 +22,13 @@ def login_required(f):
             return redirect('/')
 
     return wrap
+
+def start_session(user):
+    """Create session containing user info"""
+    del user['password']
+    session['logged_in'] = True
+    session['user'] = user
+    return jsonify(user), 200
 
 # Views
 @app.route("/")
@@ -38,7 +46,7 @@ def transcripts_view():
 @app.route("/login")
 def login_view():
     """Display log in page"""
-    return render_template("signIn.html")
+    return render_template("logIn.html")
 
 @app.route("/signup")
 def signup_view():
@@ -46,6 +54,7 @@ def signup_view():
     return render_template("signUp.html")
 
 # Form handlers
+
 @app.route("/api/upload_audio", methods=["POST"])
 def upload_audio():
     """upload audio"""
@@ -54,6 +63,48 @@ def upload_audio():
         "http://mlclient:5000/upload", files={"audio": audio_file}, timeout=5
     )
     return response.content, response.status_code
+
+@app.route('/user/signup', methods=['POST'])
+def signup():
+    """sign up"""
+    print(request.form)
+
+    # Create the user object
+    user = {
+        "username": request.form.get('username'),
+        "password": request.form.get('password')
+    }
+
+    # Encrypt the password
+    user['password'] = pbkdf2_sha256.encrypt(user['password'])
+
+    # Check for existing username address
+    if db.users.find_one({ "usernmae": user['username'] }):
+        return jsonify({ "error": "Username already in use" }), 400
+
+    if db.users.insert_one(user):
+        return start_session(user)
+
+    return jsonify({ "error": "Signup failed" }), 400
+
+@app.route('/user/signout', methods=['POST'])
+def signout():
+    """signing out"""
+    session.clear()
+    return redirect('/')
+
+@app.route('/user/login', methods=['POST'])
+def login():
+    """login"""
+    user = db.users.find_one({
+        "username": request.form.get('username')
+    })
+
+    if user and pbkdf2_sha256.verify(request.form.get('password'), user['password']):
+        return start_session(user)
+
+    return jsonify({ "error": "Invalid login credentials" }), 401
+
 
 
 if __name__ == "__main__":
