+++ web-app/templates/logIn.html
@@ -4,8 +4,11 @@
 <div class="upload-container">
     <h1>Sign In</h1>
     <form method="POST" action="/user/login" >
+        <h2>Username</h2>
         <input type="text" name="username" placeholder="username" />
+        <h2>Password</h2>
         <input type="password" name="password"/>
+        <div></div>
         <button type="submit" class="submit-btn">Log In</button>
     </form>
 </div>
