+++ web-app/templates/signUp.html
@@ -2,10 +2,13 @@
 
 {% block container %}
 <div class="upload-container">
-    <h1>Sign In</h1>
+    <h1>Sign Up</h1>
     <form method="POST" action="/user/signup">
+        <h2>Username</h2>
         <input type="text" name="username" placeholder="username" />
+        <h2>Password</h2>
         <input type="password" name="password" />
+        <div></div>
         <button type="submit" class="submit-btn">Sign Up</button>
     </form>
 </div>
